# docs/api_spec.yaml
openapi: 3.0.3
info:
  title: APP_5 API
  description: API для автоматизации принятия решений венчурными инвесторами и стартапами.
  version: 1.0.0
servers:
  - url: https://api.app5.com
    description: Production API
  - url: http://localhost:8000
    description: Local API (development)
paths:
  /auth/login:
    post:
      summary: Авторизация пользователя
      description: Возвращает JWT-токен при успешном входе.
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@app5.com
                password:
                  type: string
                  example: SecurePassword123!
      responses:
        "200":
          description: Успешная авторизация
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsIn..."
                  token_type:
                    type: string
                    example: "bearer"
        "401":
          description: Неверный email или пароль
  /auth/register:
    post:
      summary: Регистрация нового пользователя
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@app5.com
                password:
                  type: string
                  example: SecurePassword123!
                role:
                  type: string
                  enum: [investor, founder]
                  example: investor
      responses:
        "201":
          description: Пользователь успешно зарегистрирован
        "400":
          description: Ошибка валидации
...